# web/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Rempart DNS",
    "version": "1.1",
    "description": "API REST de Rempart. Session (cookie rempart_session + en-tête X-Rempart: 1 sur les requêtes qui modifient) ou jeton porteur rmp_…. Les champs inconnus sont refusés (400)."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/lists": {
      "get": {
        "tags": [
          "Listes de blocage"
        ],
        "summary": "Listes et leur état",
        "description": "Portée : read — Consultation. Tout jeton (sauf backup/sync seuls) et tout rôle.",
        "operationId": "get_api_lists",
        "x-rempart-scope": "read",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "url": {
                        "type": "string"
                      },
                      "allow": {
                        "type": "boolean"
                      },
                      "enabled": {
                        "type": "boolean"
                      },
                      "count": {
                        "type": "integer"
                      },
                      "last_update": {
                        "type": "string"
                      }
                    }
                  }
                },
                "example": [
                  {
                    "id": "x1",
                    "name": "HaGeZi Pro",
                    "url": "https://…",
                    "allow": false,
                    "enabled": true,
                    "count": 180000,
                    "last_update": "2026-10-06T03:00:00Z"
                  }
                ]
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      },
      "post": {
        "tags": [
          "Listes de blocage"
        ],
        "summary": "Ajouter une liste distante",
        "description": "https:// seulement. sha256 (facultatif) épingle le contenu attendu. allow = liste d'exceptions.\n\nPortée : admin — Modification de la configuration. Jeton « admin », rôle opérateur ou administrateur.",
        "operationId": "post_api_lists",
        "x-rempart-scope": "admin",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "url": {
                    "type": "string"
                  },
                  "sha256": {
                    "type": "string"
                  },
                  "allow": {
                    "type": "boolean"
                  }
                }
              },
              "example": {
                "name": "Ma liste",
                "url": "https://exemple.fr/liste.txt",
                "sha256": "",
                "allow": false
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Jeton rmp_… créé dans Sécurité → Accès API ou par POST /api/tokens."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rempart_session",
        "description": "Cookie posé par POST /api/login."
      }
    },
    "responses": {
      "Error": {
        "description": "Erreur",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}